+++ apps/eternal-devotion/src/app/components/beautiful-body/index.ts
@@ -1,1 +1,3 @@
-export { default } from './beautiful_body_layout';
+export { default as BeautifulBodyLayout } from './beautiful_body_layout';
+export { default as SliceIndicator } from './slice_indicator';
+export { randomizeSlots } from './config';
