+++ apps/eternal-devotion/src/app/components/beautiful-body/slice_indicator.module.scss
@@ -1,0 +1,46 @@
+// Adapted from the offering page's position_indicator (see
+// offering/position_indicator.module.scss): same size, font, anchor, and
+// escalating padding-top stagger. Spacing uses a flex column gap here rather
+// than per-span 1ch widths, and renders three slice values divided by slashes
+// (value / value / value → 5 spans) instead of current / total.
+.indicator {
+  --indicator--size: 10dvw;
+  --indicator--stagger: 0.25;
+
+  position: absolute;
+  top: 0;
+  left: var(--padding--page);
+  font-size: var(--indicator--size);
+  font-family: var(--font--serif);
+  font-style: italic;
+  display: flex;
+  column-gap: 0.5ch;
+  pointer-events: none;
+
+  span {
+    line-height: 0.85;
+
+    // Even 0.5x steps across all five spans (3 values + 2 slashes) so the
+    // cascade stays contained near the top rather than trailing off — the
+    // offering page only staggers 4 spans, we have 5.
+    &:nth-child(2) {
+      padding-top: calc(
+        var(--indicator--size) * var(--indicator--stagger) * 0.5
+      );
+    }
+
+    &:nth-child(3) {
+      padding-top: calc(var(--indicator--size) * var(--indicator--stagger));
+    }
+
+    &:nth-child(4) {
+      padding-top: calc(
+        var(--indicator--size) * var(--indicator--stagger) * 1.5
+      );
+    }
+
+    &:nth-child(5) {
+      padding-top: calc(var(--indicator--size) * var(--indicator--stagger) * 2);
+    }
+  }
+}
