+++ apps/eternal-devotion/src/app/components/offering_selection/selection_layout.tsx
@@ -1,8 +1,22 @@
 import { createContentPathname, offeringsConfig } from 'content';
+import type { OfferingsGroup } from 'content';
+import { createOfferingPath } from 'utils';
 
 import SelectionCard from './selection_card';
 import SelectionCarousel from './selection_carousel';
 import styles from './selection_layout.module.scss';
+
+// Where a card links, by offering kind. Paintings open the first offering;
+// experiences go to their own route.
+const getCardPath = (group: OfferingsGroup) =>
+  group.kind === 'experience' ? group.path : createOfferingPath(group.id, '1');
+
+// Card cover image, by offering kind. Both kinds have a root-level cover keyed
+// by filename; painting covers are .jpg, the experience cover is .png.
+const getCardImageUrl = (group: OfferingsGroup) =>
+  group.kind === 'experience'
+    ? createContentPathname(`${group.filename}.png`, 'root')
+    : createContentPathname(`${group.filename}.jpg`, 'root');
 
 const SelectionLayout = () => {
   return (
@@ -18,11 +32,11 @@ const SelectionLayout = () => {
           <SelectionCard
             callToAction={group.callToAction}
             descriptionList={group.descriptionList}
-            id={groupId}
-            imageUrl={createContentPathname(`${group.filename}.jpg`, 'root')}
+            imageUrl={getCardImageUrl(group)}
             index={index}
             key={groupId}
             title={group.title}
+            to={getCardPath(group)}
           />
         ))}
       </SelectionCarousel>
