+++ apps/eternal-devotion/src/app/routes/offering.tsx
@@ -1,7 +1,7 @@
 import { Navigate } from 'react-router';
 
 import { useCurrentOffering } from 'content';
-import { createOfferingPath } from 'utils';
+import { getSelectionPath } from 'utils';
 import { PaintingRow, PositionIndicator } from 'components/offering';
 
 import styles from './offering.module.scss';
@@ -9,9 +9,9 @@ import styles from './offering.module.scss';
 const Offering = () => {
   const currentOffering = useCurrentOffering();
 
-  // You've fallen astray
+  // You've fallen astray — an unknown group/id lands back at the selection.
   if (!currentOffering) {
-    return <Navigate replace to={createOfferingPath('1', '1')} />;
+    return <Navigate replace to={getSelectionPath()} />;
   }
 
   return (
