+++ apps/eternal-devotion/src/content/index.ts
@@ -3,3 +3,11 @@ import offeringsConfig, { IMAGE_SIZES } from './config';
 export { offeringsConfig, IMAGE_SIZES };
 export * from './types';
 export * from './utils';
+
+// Beautiful-body (exquisite corpse) content.
+export {
+  CHARACTERS,
+  PIECES,
+  BeautifulBodyGroup
+} from './offerings/beautiful_body';
+export type { Character } from './offerings/beautiful_body';
